Tie New to a particular vector type
We now have (New v a) which can only be used to create vectors of type (v a)
where before, (New a) could be used to create arbitrary vectors. This allows
the initialiser in New to include operations which only work for a particular
implementation of Vector. This will be quite useful in the future, I'm sure.