Allow streams to produce entire vectors as well as individual elements
This is a signficant extension to the stream fusion mechanism. A Stream is now
parametrised with a vector type and offers two iteration modes: individual
elements and individual elements + vectors. This supports significantly more
efficient implementations of block operations. For instance, (++) will now use
memcpy where possible rather than iterating through the elements in a loop.

Comments are sorely lacking at the moment.