Modify GenUnboxTuple to not generate warnings
Many definitions ignore some of their arguments, and with -Wall enabled, these
will be warned about unless the variable names begin with an underscore. So the
code generator has been modified to put leading underscores on most variable
names. These are still legal variable names to refer to (they don't act like _,
which is a special pattern, not a variable name). This successfully makes the
generated code compile warning-free, although it isn't identical to the code
that is checked in.