Fix dropWhile performance
The `unstream` version results in a copy of the remainder of the vector
which is rarely what you want to do. This commit uses `drop` directly
after `findIndex` which results in a big speed gain if the remaining
part of the vector is large.

Cf. https://github.com/haskell/vector/issues/108
and https://github.com/haskell/vector/issues/141