Add since pragmas to every function added since last release
It's not clear what is right way to add since pragmas. Their values are decided
when release is made and not when function is added.

One possibility is to write since NEXT and then sed it to correct value when
doing release. One question is how to not forget it