Add GHC 9.0 to CI
Turns out it requires cabal 3.4 to build and it's only available on ubuntu ATM
so no 9.0 on macOS and windows. Hopefully this will be fixed at some point.

Doctests don't work on GHC 9.0 as well as on 8.10. We need to fix that

Also bump dependency on doctests. Things seem to work well.