Drop functions which don't enforce strictness
These functions require user to maintain invariant that all element of vector is
in WHNF. And it's well known that this insvariant is difficult to enforce and
its violations are usually invisible.

Such spec leaks may be invisibe during testing only to cause pproblem in
production. So it's better to remove them unless good use case for them appears