Enable CPP per module in tests
This is workaround for problem of benchmarking build time.
When package is compiled with GHC flags `-ddump-timings -ddump-to-file`
*.dump-timings files corresponding to modules with enabled CPP gets overwritten
at link time.

Timing for Utilities still get lost but its contribution is small anyway
and we're able to keep timings for other modules

Note that libraries are not affected by this problem.